refactor(benchmarks-website): split migrate.rs into migrate/ module
Pull the four `*Accum` structs and their `build_*_batch` functions into
a sibling `migrate/accum.rs`, leaving `migrate/mod.rs` as the
orchestrator (`MigrationSummary`, `open_target_db`, `run`, `flush_all`,
`apply_v2_record`, `migrate_data_jsonl`, `migrate_file_sizes`, `flush`).

The four accumulators are kept as siblings rather than abstracted
behind a trait. They share the dedup-on-`measurement_id` shape but
their parallel-vec layouts diverge (CompressionSize uses a HashMap and
has two collision semantics — replace and sum), and the existing
push/build code is more legible when each accumulator is a separate,
self-contained type.

Behaviour preserved: every batch schema, dedup semantic, and Display
output stays identical.
