+++ benchmarks-website/migrate/src/migrate/mod.rs
@@ -4,34 +4,27 @@
 //! End-to-end migration of one v2 dataset into a v3 DuckDB file.
 //!
 //! Streams `data.json.gz` line-by-line, runs each record through the
-//! [`classifier`], and writes one row per record into the appropriate v3 fact table.
-//! Every row's `measurement_id` is computed via the server's `measurement_id_*` functions so the
-//! result is byte-compatible with what fresh `/api/ingest` would have produced.
+//! [`crate::classifier`], and writes one row per record into the appropriate
+//! v3 fact table. Every row's `measurement_id` is computed via the server's
+//! `measurement_id_*` functions so the result is byte-compatible with what
+//! fresh `/api/ingest` would have produced.
 //!
 //! Bulk-load shape: rows are accumulated in memory as parallel column
 //! vectors, deduplicated by `measurement_id`, then flushed to DuckDB
 //! via `Appender::append_record_batch` as one Arrow `RecordBatch` per
 //! fact table.
 
+mod accum;
+
 use std::collections::BTreeMap;
 use std::io::BufRead;
 use std::path::Path;
-use std::sync::Arc;
 use std::time::Duration;
 use std::time::Instant;
 
 use anyhow::Context as _;
 use anyhow::Result;
-use arrow_array::ArrayRef;
-use arrow_array::Int32Array;
-use arrow_array::Int64Array;
-use arrow_array::ListArray;
 use arrow_array::RecordBatch;
-use arrow_array::StringArray;
-use arrow_buffer::OffsetBuffer;
-use arrow_schema::DataType;
-use arrow_schema::Field;
-use arrow_schema::Schema;
 use duckdb::Connection;
 use tracing::info;
 use tracing::warn;
@@ -44,8 +37,15 @@ use vortex_bench_server::records::CompressionTime;
 use vortex_bench_server::records::QueryMeasurement;
 use vortex_bench_server::records::RandomAccessTime;
 use vortex_bench_server::schema::SCHEMA_DDL;
-use vortex_utils::aliases::hash_map::HashMap;
 
+use self::accum::CompressionSizeAccum;
+use self::accum::CompressionTimeAccum;
+use self::accum::QueryAccum;
+use self::accum::RandomAccessAccum;
+use self::accum::build_compression_size_batch;
+use self::accum::build_compression_time_batch;
+use self::accum::build_query_batch;
+use self::accum::build_random_access_batch;
 use crate::classifier;
 use crate::classifier::V3Bin;
 use crate::commits::upsert_commit;
@@ -62,19 +62,34 @@ use crate::v2::value_as_f64;
 /// Per-table insert counts, plus skip / missing counts.
 #[derive(Debug, Default, Clone)]
 pub struct MigrationSummary {
+    /// Lines read from `data.json.gz`.
     pub records_read: u64,
+    /// Rows successfully inserted into `query_measurements`.
     pub query_inserted: u64,
+    /// Rows successfully inserted into `compression_times`.
     pub compression_time_inserted: u64,
+    /// Rows successfully inserted into `compression_sizes`.
     pub compression_size_inserted: u64,
+    /// Rows successfully inserted into `random_access_times`.
     pub random_access_inserted: u64,
+    /// `file-sizes-*.json.gz` lines folded into `compression_sizes`.
     pub file_size_inserted: u64,
+    /// Records the classifier returned `Unknown` for.
     pub uncategorized: u64,
+    /// Top-level prefix histogram of uncategorised records, for triage.
     pub uncategorized_prefixes: BTreeMap<String, u64>,
+    /// Records whose `commit_id` doesn't match any commit in `commits.jsonl`.
     pub missing_commit: u64,
+    /// Warnings emitted while upserting commits (e.g. missing tree SHA).
     pub commit_warnings: u64,
+    /// Records dropped because their `value` was missing or non-numeric.
     pub skipped_no_value: u64,
+    /// Records the classifier returned `Skip(reason)` for.
     pub skipped_intentional: u64,
+    /// Commits upserted into the `commits` dim table.
     pub commits_inserted: u64,
+    /// Records dropped by dedup because their `measurement_id` collided
+    /// with a previously kept row.
     pub deduped: u64,
     /// Number of records dropped by dedup whose `value_ns` (or
     /// `value_bytes` for compression_sizes' replace path) differed
@@ -505,320 +520,6 @@ fn flush(conn: &Connection, table: &str, batch: RecordBatch) -> Result<()> {
         .with_context(|| format!("appending record batch to {table}"))?;
     drop(app);
     Ok(())
-}
-
-#[derive(Default)]
-struct QueryAccum {
-    measurement_id: Vec<i64>,
-    commit_sha: Vec<String>,
-    dataset: Vec<String>,
-    dataset_variant: Vec<Option<String>>,
-    scale_factor: Vec<Option<String>>,
-    query_idx: Vec<i32>,
-    storage: Vec<String>,
-    engine: Vec<String>,
-    format: Vec<String>,
-    value_ns: Vec<i64>,
-    all_runtimes_ns: Vec<Vec<i64>>,
-    peak_physical: Vec<Option<i64>>,
-    peak_virtual: Vec<Option<i64>>,
-    physical_delta: Vec<Option<i64>>,
-    virtual_delta: Vec<Option<i64>>,
-    env_triple: Vec<Option<String>>,
-    /// `mid` -> index in the parallel column vecs. Lets us look up the
-    /// kept row's `value_ns` on collision so we can flag conflicts.
-    seen: HashMap<i64, usize>,
-}
-
-impl QueryAccum {
-    fn push(&mut self, mid: i64, r: QueryMeasurement, summary: &mut MigrationSummary) {
-        if let Some(&idx) = self.seen.get(&mid) {
-            summary.deduped += 1;
-            if self.value_ns[idx] != r.value_ns {
-                summary.deduped_with_conflict += 1;
-            }
-            return;
-        }
-        let idx = self.measurement_id.len();
-        self.seen.insert(mid, idx);
-        self.measurement_id.push(mid);
-        self.commit_sha.push(r.commit_sha);
-        self.dataset.push(r.dataset);
-        self.dataset_variant.push(r.dataset_variant);
-        self.scale_factor.push(r.scale_factor);
-        self.query_idx.push(r.query_idx);
-        self.storage.push(r.storage);
-        self.engine.push(r.engine);
-        self.format.push(r.format);
-        self.value_ns.push(r.value_ns);
-        self.all_runtimes_ns.push(r.all_runtimes_ns);
-        self.peak_physical.push(r.peak_physical);
-        self.peak_virtual.push(r.peak_virtual);
-        self.physical_delta.push(r.physical_delta);
-        self.virtual_delta.push(r.virtual_delta);
-        self.env_triple.push(r.env_triple);
-    }
-}
-
-#[derive(Default)]
-struct CompressionTimeAccum {
-    measurement_id: Vec<i64>,
-    commit_sha: Vec<String>,
-    dataset: Vec<String>,
-    dataset_variant: Vec<Option<String>>,
-    format: Vec<String>,
-    op: Vec<String>,
-    value_ns: Vec<i64>,
-    all_runtimes_ns: Vec<Vec<i64>>,
-    env_triple: Vec<Option<String>>,
-    seen: HashMap<i64, usize>,
-}
-
-impl CompressionTimeAccum {
-    fn push(&mut self, mid: i64, r: CompressionTime, summary: &mut MigrationSummary) {
-        if let Some(&idx) = self.seen.get(&mid) {
-            summary.deduped += 1;
-            if self.value_ns[idx] != r.value_ns {
-                summary.deduped_with_conflict += 1;
-            }
-            return;
-        }
-        let idx = self.measurement_id.len();
-        self.seen.insert(mid, idx);
-        self.measurement_id.push(mid);
-        self.commit_sha.push(r.commit_sha);
-        self.dataset.push(r.dataset);
-        self.dataset_variant.push(r.dataset_variant);
-        self.format.push(r.format);
-        self.op.push(r.op);
-        self.value_ns.push(r.value_ns);
-        self.all_runtimes_ns.push(r.all_runtimes_ns);
-        self.env_triple.push(r.env_triple);
-    }
-}
-
-#[derive(Default)]
-struct RandomAccessAccum {
-    measurement_id: Vec<i64>,
-    commit_sha: Vec<String>,
-    dataset: Vec<String>,
-    format: Vec<String>,
-    value_ns: Vec<i64>,
-    all_runtimes_ns: Vec<Vec<i64>>,
-    env_triple: Vec<Option<String>>,
-    seen: HashMap<i64, usize>,
-}
-
-impl RandomAccessAccum {
-    fn push(&mut self, mid: i64, r: RandomAccessTime, summary: &mut MigrationSummary) {
-        if let Some(&idx) = self.seen.get(&mid) {
-            summary.deduped += 1;
-            if self.value_ns[idx] != r.value_ns {
-                summary.deduped_with_conflict += 1;
-            }
-            return;
-        }
-        let idx = self.measurement_id.len();
-        self.seen.insert(mid, idx);
-        self.measurement_id.push(mid);
-        self.commit_sha.push(r.commit_sha);
-        self.dataset.push(r.dataset);
-        self.format.push(r.format);
-        self.value_ns.push(r.value_ns);
-        self.all_runtimes_ns.push(r.all_runtimes_ns);
-        self.env_triple.push(r.env_triple);
-    }
-}
-
-/// `compression_sizes` is fed by both data.json.gz (replace-on-collision)
-/// and file-sizes-*.json.gz (sum-on-collision). Stored as a map; converted
-/// to a `RecordBatch` at flush time.
-#[derive(Default)]
-struct CompressionSizeAccum {
-    rows: HashMap<i64, CompressionSize>,
-}
-
-impl CompressionSizeAccum {
-    /// data.json.gz path: latest write wins, mirroring the prior
-    /// `ON CONFLICT DO UPDATE SET value_bytes = excluded.value_bytes`.
-    /// Bumps `deduped_with_conflict` when an existing row's
-    /// `value_bytes` differs from the incoming row's, so silent
-    /// value-corruption is observable.
-    fn push_replace(&mut self, mid: i64, r: CompressionSize, summary: &mut MigrationSummary) {
-        if let Some(existing) = self.rows.get(&mid)
-            && existing.value_bytes != r.value_bytes
-        {
-            summary.deduped_with_conflict += 1;
-        }
-        self.rows.insert(mid, r);
-    }
-
-    /// file-sizes-*.json.gz path: per-file rows aggregate into one
-    /// `(commit, dataset, dataset_variant, format)` row by summing,
-    /// mirroring the prior `value_bytes = compression_sizes.value_bytes
-    /// + excluded.value_bytes`.
-    fn push_sum(&mut self, mid: i64, r: CompressionSize) {
-        let add = r.value_bytes;
-        self.rows
-            .entry(mid)
-            .and_modify(|x| x.value_bytes += add)
-            .or_insert(r);
-    }
-}
-
-fn build_query_batch(a: QueryAccum) -> Result<RecordBatch> {
-    let schema = Arc::new(Schema::new(vec![
-        Field::new("measurement_id", DataType::Int64, false),
-        Field::new("commit_sha", DataType::Utf8, false),
-        Field::new("dataset", DataType::Utf8, false),
-        Field::new("dataset_variant", DataType::Utf8, true),
-        Field::new("scale_factor", DataType::Utf8, true),
-        Field::new("query_idx", DataType::Int32, false),
-        Field::new("storage", DataType::Utf8, false),
-        Field::new("engine", DataType::Utf8, false),
-        Field::new("format", DataType::Utf8, false),
-        Field::new("value_ns", DataType::Int64, false),
-        Field::new(
-            "all_runtimes_ns",
-            DataType::List(Arc::new(Field::new("item", DataType::Int64, false))),
-            false,
-        ),
-        Field::new("peak_physical", DataType::Int64, true),
-        Field::new("peak_virtual", DataType::Int64, true),
-        Field::new("physical_delta", DataType::Int64, true),
-        Field::new("virtual_delta", DataType::Int64, true),
-        Field::new("env_triple", DataType::Utf8, true),
-    ]));
-    let cols: Vec<ArrayRef> = vec![
-        Arc::new(Int64Array::from(a.measurement_id)),
-        Arc::new(StringArray::from(a.commit_sha)),
-        Arc::new(StringArray::from(a.dataset)),
-        Arc::new(StringArray::from(a.dataset_variant)),
-        Arc::new(StringArray::from(a.scale_factor)),
-        Arc::new(Int32Array::from(a.query_idx)),
-        Arc::new(StringArray::from(a.storage)),
-        Arc::new(StringArray::from(a.engine)),
-        Arc::new(StringArray::from(a.format)),
-        Arc::new(Int64Array::from(a.value_ns)),
-        Arc::new(build_list_int64(a.all_runtimes_ns)),
-        Arc::new(Int64Array::from(a.peak_physical)),
-        Arc::new(Int64Array::from(a.peak_virtual)),
-        Arc::new(Int64Array::from(a.physical_delta)),
-        Arc::new(Int64Array::from(a.virtual_delta)),
-        Arc::new(StringArray::from(a.env_triple)),
-    ];
-    Ok(RecordBatch::try_new(schema, cols)?)
-}
-
-fn build_compression_time_batch(a: CompressionTimeAccum) -> Result<RecordBatch> {
-    let schema = Arc::new(Schema::new(vec![
-        Field::new("measurement_id", DataType::Int64, false),
-        Field::new("commit_sha", DataType::Utf8, false),
-        Field::new("dataset", DataType::Utf8, false),
-        Field::new("dataset_variant", DataType::Utf8, true),
-        Field::new("format", DataType::Utf8, false),
-        Field::new("op", DataType::Utf8, false),
-        Field::new("value_ns", DataType::Int64, false),
-        Field::new(
-            "all_runtimes_ns",
-            DataType::List(Arc::new(Field::new("item", DataType::Int64, false))),
-            false,
-        ),
-        Field::new("env_triple", DataType::Utf8, true),
-    ]));
-    let cols: Vec<ArrayRef> = vec![
-        Arc::new(Int64Array::from(a.measurement_id)),
-        Arc::new(StringArray::from(a.commit_sha)),
-        Arc::new(StringArray::from(a.dataset)),
-        Arc::new(StringArray::from(a.dataset_variant)),
-        Arc::new(StringArray::from(a.format)),
-        Arc::new(StringArray::from(a.op)),
-        Arc::new(Int64Array::from(a.value_ns)),
-        Arc::new(build_list_int64(a.all_runtimes_ns)),
-        Arc::new(StringArray::from(a.env_triple)),
-    ];
-    Ok(RecordBatch::try_new(schema, cols)?)
-}
-
-fn build_random_access_batch(a: RandomAccessAccum) -> Result<RecordBatch> {
-    let schema = Arc::new(Schema::new(vec![
-        Field::new("measurement_id", DataType::Int64, false),
-        Field::new("commit_sha", DataType::Utf8, false),
-        Field::new("dataset", DataType::Utf8, false),
-        Field::new("format", DataType::Utf8, false),
-        Field::new("value_ns", DataType::Int64, false),
-        Field::new(
-            "all_runtimes_ns",
-            DataType::List(Arc::new(Field::new("item", DataType::Int64, false))),
-            false,
-        ),
-        Field::new("env_triple", DataType::Utf8, true),
-    ]));
-    let cols: Vec<ArrayRef> = vec![
-        Arc::new(Int64Array::from(a.measurement_id)),
-        Arc::new(StringArray::from(a.commit_sha)),
-        Arc::new(StringArray::from(a.dataset)),
-        Arc::new(StringArray::from(a.format)),
-        Arc::new(Int64Array::from(a.value_ns)),
-        Arc::new(build_list_int64(a.all_runtimes_ns)),
-        Arc::new(StringArray::from(a.env_triple)),
-    ];
-    Ok(RecordBatch::try_new(schema, cols)?)
-}
-
-fn build_compression_size_batch(a: CompressionSizeAccum) -> Result<RecordBatch> {
-    let n = a.rows.len();
-    let mut measurement_id = Vec::with_capacity(n);
-    let mut commit_sha = Vec::with_capacity(n);
-    let mut dataset = Vec::with_capacity(n);
-    let mut dataset_variant = Vec::with_capacity(n);
-    let mut format = Vec::with_capacity(n);
-    let mut value_bytes = Vec::with_capacity(n);
-    for (mid, cs) in a.rows {
-        measurement_id.push(mid);
-        commit_sha.push(cs.commit_sha);
-        dataset.push(cs.dataset);
-        dataset_variant.push(cs.dataset_variant);
-        format.push(cs.format);
-        value_bytes.push(cs.value_bytes);
-    }
-    let schema = Arc::new(Schema::new(vec![
-        Field::new("measurement_id", DataType::Int64, false),
-        Field::new("commit_sha", DataType::Utf8, false),
-        Field::new("dataset", DataType::Utf8, false),
-        Field::new("dataset_variant", DataType::Utf8, true),
-        Field::new("format", DataType::Utf8, false),
-        Field::new("value_bytes", DataType::Int64, false),
-    ]));
-    let cols: Vec<ArrayRef> = vec![
-        Arc::new(Int64Array::from(measurement_id)),
-        Arc::new(StringArray::from(commit_sha)),
-        Arc::new(StringArray::from(dataset)),
-        Arc::new(StringArray::from(dataset_variant)),
-        Arc::new(StringArray::from(format)),
-        Arc::new(Int64Array::from(value_bytes)),
-    ];
-    Ok(RecordBatch::try_new(schema, cols)?)
-}
-
-/// Build a non-nullable `List<Int64>` Arrow array from one inner Vec
-/// per row. The outer list is non-null; inner i64 values are non-null.
-fn build_list_int64(values: Vec<Vec<i64>>) -> ListArray {
-    let mut offsets: Vec<i32> = Vec::with_capacity(values.len() + 1);
-    offsets.push(0);
-    let mut flat: Vec<i64> = Vec::new();
-    for inner in values {
-        flat.extend_from_slice(&inner);
-        offsets.push(flat.len() as i32);
-    }
-    let values_arr = Int64Array::from(flat);
-    let field = Arc::new(Field::new("item", DataType::Int64, false));
-    ListArray::new(
-        field,
-        OffsetBuffer::new(offsets.into()),
-        Arc::new(values_arr),
-        None,
-    )
 }
 
 /// Print the summary in a human-readable form. Returned by the CLI.
